+++ src/lib/index.ts
@@ -1,9 +1,0 @@
-import PushReceiver from "../pushReceiver"
-import PushSender from "../pushSender"
-
-export {
-    PushReceiver,
-    PushSender,
-}
-
-export default PushReceiver